Renamed the package name from config_server and discovery_service to configserver and discoveryservice to follow the standard naming conventions.
