+++ services/config-server/src/main/java/com/example/ecommerce/configserver/ConfigServerApplication.java
@@ -1,4 +1,4 @@
-package com.example.ecommerce.config_server;
+package com.example.ecommerce.configserver;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
