+++ services/discovery-service/src/main/java/com/example/ecommerce/discoveryservice/DiscoveryServiceApplication.java
@@ -1,4 +1,4 @@
-package com.example.ecommerce.discovery_service;
+package com.example.ecommerce.discoveryservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
