+++ services/discovery-service/src/test/java/com/example/ecommerce/discoveryservice/DiscoveryServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.ecommerce.discovery_service;
+package com.example.ecommerce.discoveryservice;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
